Update CustomGodot dialogs
As part of the Change Wording Project, messages have been updated to be more similar to that of Godot.
Error messages will also be shown when you attempt to use a non-existent file or a invalid executable.

For macOS: automatically detect C# support through checking if the tag has "mono".

BrowseGodotDialog's default path is now set to EnginePath per use. Finally, the "Is Mono" checkbox has been removed.